Use JSON manifests and introduce version_manifest
While it should've been done in two steps, I did it in one. This patch
moves away from XML manifests to JSON ones and moves the concerns of
parsing and validating the JSON manifest(s) to a version_manifest class.

The revised implementation also moves away from caching certain things
in object state like the current version, which releases were applied,
etc. Although this degrades performance (particularly where we have to
re-compute the identity list checksums) it will prove for a saner and
more reliable API.

The version_manifest exposes an API for querying the current version of
the application and the list of available releases to be applied which
in turn the patcher can consume for applying an update.