Stop removing empty directories in op_create
So we were tracking whether the destination directory existed or not and
whether we were responsible for its creation during staging so that in
rollback we would remove those directories. The problem was that this
behavior is not predictable nor idempotent; if staging and rollback were
done in different processes, the outcome is not consistent.

This patch removes that behavior altogether so that we achieve
consistency although at the cost of some (harmless) side-effects. Empty
directories aren't a big deal to fret so much about.